+++ src/main/java/com/ruvalcaba/quizapplication/DAO/QuizDAO.java
@@ -3,5 +3,5 @@ package com.ruvalcaba.quizapplication.DAO;
 import com.ruvalcaba.quizapplication.model.QuizModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuizDAO extends JpaRepository<QuizModel, Integer> {
+public interface QuizDAO extends JpaRepository<QuizModel, Long> {
 }
